+++ src/config/rt.rs
@@ -11,6 +11,8 @@ use crate::config::{
     ConfigOptsTools, ConfigOptsWatch,
 };
 
+use super::models::WsProtocol;
+
 /// Config options for the cargo build command
 #[derive(Clone, Debug)]
 pub enum Features {
@@ -218,6 +220,8 @@ pub struct RtcServe {
     pub proxies: Option<Vec<ConfigOptsProxy>>,
     /// Whether to disable auto-reload of the web page when a build completes.
     pub no_autoreload: bool,
+    /// Protocol used for autoreload WebSockets connection.
+    pub ws_protocol: Option<WsProtocol>,
 }
 
 impl RtcServe {
@@ -247,6 +251,7 @@ impl RtcServe {
             proxy_ws: opts.proxy_ws,
             proxies,
             no_autoreload: opts.no_autoreload,
+            ws_protocol: opts.ws_protocol,
         })
     }
 }